fix: Performance and hardware column sorting
