+++ packages/frontend/src/components/List/index.ts
@@ -1,3 +1,4 @@
+export * from './Column';
 export * from './List';
 export * from './Truncate';
 export * from './Row';
